[#1] AG | Deploy concourse to full BOSH AWS cluster
Motivation
==========

__BUG:__ Github authorization fails on bosh-lite platform.

__CAUSE:__ The bosh-lite installation of concourse makes the concourse
server available on an "external" ip address, that is actually an
internal address (`10.244.8.2:8080`) within a cluster of Warden
running on one bosh-lite VM. But github requires that the URL to which
it sends an OAuth token callback to have the same base URL as the server
that originated it. Because the originating URL is `10.244.8.2:8080`,
but the target (externally reachable URL) is `ci.whereat.io` (or the
VPC's elastic IP address -- no difference to the bug), Github refuses
the OAuth request. If you try to provide a different value for the
`external_url` in the concourse `yml` manifest, it breaks the deploy. Grr..

__FIX_: Figure out deploying to a full-on BOSH cluster, which we
want for pipelines that incorporate Mac servers down the line anyway!

Commit Notes
============

* Create new file structure to separate `bosh`, `concourse`, and
  `concourse-lite` con figs
* Document lessons about configuring Bosh/Concourse/AWS in `README`
* Use an `insert-secrets.sh` script so we can keep one manifest under
  version control instead of separate `dist` and live copies, thusly:
  * only `dist` manifests are kept under version control, with
  placeholder values for secrets
  * at runtime, script creates a live copy of manifest and replaces all
  secrete placeholders with values from environment variables
  * after a successful deployment, live manifest is deleted
  * after a failed deployment, live manifest is left on disk so
  developer may inspect it for errors